refactor(mobileapp): create auth folder in screens folder
