+++ mobileapp/src/screens/auth/ForgotPassword.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import SimpleLayout from "../components/layouts/SimpleLayout";
-import { Heading } from "../components/atoms/Text";
+import SimpleLayout from "../../components/layouts/SimpleLayout";
+import { Heading } from "../../components/atoms/Text";
 
 export default function ForgotPassword() {
   return (
